Include query string in HTTP request line
The SimpleHttpClient was only sending the path portion of the URI
in the HTTP request line, omitting the query string. This caused
the server to not receive query parameters.

HTTP/1.1 requires the full request-target including query string:
  GET /search?q=test&page=1 HTTP/1.1

Modified sendRequest to build pathWithQuery that includes the
query component when present using Option to avoid null checks.

Fixes query parameters test.